// Project_tree/myFamilyTree/src/models/tintuc.ts
export interface TinTuc {
  tinTucId?: string;
  dongHoId: string;
  tieuDe: string;
  noiDung?: string;
  tomTat?: string;
  anhDaiDien?: string;
  tacGia?: string;
  ngayDang?: Date;
  luotXem?: number;
  ghim?: number;
  active_flag?: number;
  nguoiTaoId?: string;
  lu_updated?: Date;
  lu_user_id?: string;
  tenDongHo?: string;
  RecordCount?: number;
}